+++ nestjs-api/src/pages/roles/role.controller.ts
@@ -9,6 +9,7 @@ import {
   Query,
   ParseIntPipe,
   Req,
+  ForbiddenException,
 } from '@nestjs/common';
 import { RolesService } from './role.service';
 import { CreateDto, UpdateDto } from './dto/role.dto';
@@ -16,11 +17,24 @@ import { RoleEntity } from './entity/role.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Request } from 'express';
+import { AppointmentStatusEntity } from '../appointmentstatuses/entity/appointmentstatus.entity';
 
 import { getQueryData } from '../../lib/functions';
 import { RoleAccessDetailsService } from '../roleaccessdetails/roleaccessdetail.service';
 import { RoleAccessDetailEntity } from '../roleaccessdetails/entity/roleaccessdetail.entity';
 import { In } from 'typeorm';
+import {
+  ValidateAccessMethod,
+  isGranted,
+  getValidationError,
+} from 'src/component/validateaccess/validate-access.decorator';
+import {
+  denyRoleBasedAccess,
+  hasCreateAccess,
+  hasDeleteAccess,
+  hasReadAccess,
+  hasUpdateAccess,
+} from 'src/component/validateaccess/validate-rbactoken';
 
 interface PaginationQuery {
   page?: number;
@@ -32,39 +46,78 @@ export class RolesController {
   constructor(
     private readonly rolesService: RolesService,
     private readonly roleAccessDetailsService: RoleAccessDetailsService,
-    @InjectRepository(RoleAccessDetailEntity)
-    private readonly roleAccessDetailRepository: Repository<RoleAccessDetailEntity>,
+    @InjectRepository(AppointmentStatusEntity)
+    private readonly appointmentStatusRepository: Repository<AppointmentStatusEntity>,
   ) {}
 
-  @Get('getallaccessoptions')
-  async getAllAccessOptions(
-    @Query() query: PaginationQuery,
-  ): Promise<{ data: Record<string, { value: string; label: string }[]> }> {
-    let typeid = getQueryData(query, 'typeid', [2, 3]);
-    if (!Array.isArray(typeid)) {
-      typeid = [Number(typeid)];
+  // @Get('getoption/appointmentstatusoptions')
+  // async getAppointmentStatusOptions(): Promise<{
+  //   data: { value: string; label: string }[];
+  // }> {
+  //   const result = await this.rolesService.getAppointmentStatusOptions();
+  //   console.log('result Options:', result);
+  //   return {
+  //     data: result,
+  //   };
+  // }
+
+  @Get('getoptionbyquerystring')
+  @ValidateAccessMethod({ RBACModule: 'roles' })
+  async getOptionsByModel(@Query('model') model: string, @Req() req: Request) {
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
     }
-    const typeidArray = typeid.map((t) => Number(t));
-    const data = await this.roleAccessDetailRepository.find({
-      where: { typeid: In(typeidArray) },
-    });
+    if (!hasReadAccess(req)) return denyRoleBasedAccess();
 
-    const result: Record<string, { value: string; label: string }[]> = {};
-    data.forEach((item) => {
-      if (!result[item.accesskey]) {
-        const accessArray = Array.isArray(item.access) ? item.access : [];
-        result[item.accesskey] = accessArray.map((acc: string) => ({
-          value: acc,
-          label: acc,
-        }));
-      }
-    });
-
-    return { data: result };
+    if (!model) {
+      throw new BadRequestException('Model parameter is required');
+    }
+    switch (model) {
+      case 'appointmentstatuses':
+        const result = await this.rolesService.getAppointmentStatusOptions();
+        return {
+          data: result,
+        };
+      default:
+        throw new BadRequestException(`Unsupported model: ${model}`);
+    }
   }
+  // @Get('get')
+
+  // @Get('getallaccessoptions')
+  // async getAllAccessOptions(
+  //   @Query() query: PaginationQuery,
+  // ): Promise<{ data: Record<string, { value: string; label: string }[]> }> {
+  //   let typeid = getQueryData(query, 'typeid', [2, 3]);
+  //   if (!Array.isArray(typeid)) {
+  //     typeid = [Number(typeid)];
+  //   }
+  //   const typeidArray = typeid.map((t) => Number(t));
+  //   const data = await this.roleAccessDetailRepository.find({
+  //     where: { typeid: In(typeidArray) },
+  //   });
+
+  //   const result: Record<string, { value: string; label: string }[]> = {};
+  //   data.forEach((item) => {
+  //     if (!result[item.accesskey]) {
+  //       const accessArray = Array.isArray(item.access) ? item.access : [];
+  //       result[item.accesskey] = accessArray.map((acc: string) => ({
+  //         value: acc,
+  //         label: acc,
+  //       }));
+  //     }
+  //   });
+
+  //   return { data: result };
+  // }
 
   @Get('index')
-  async findAll(@Query() query: PaginationQuery) {
+  @ValidateAccessMethod({ RBACModule: 'roles' })
+  async findAll(@Query() query: PaginationQuery, @Req() req: Request) {
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
+    }
+    if (!hasReadAccess(req)) return denyRoleBasedAccess();
     const page = Number(query.page) || 1;
     const searchCond = query.searchcond?.trim() || '';
 
@@ -72,15 +125,24 @@ export class RolesController {
   }
 
   @Get('getcard')
-  async findOne(@Query('id', ParseIntPipe) id: number) {
+  @ValidateAccessMethod({ RBACModule: 'roles' })
+  async findOne(@Query('id', ParseIntPipe) id: number, @Req() req: Request) {
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
+    }
+    if (!hasReadAccess(req)) return denyRoleBasedAccess();
     const result = await this.rolesService.findOne(RoleEntity, id);
     return { data: result };
   }
 
   @Post('newcard')
+  @ValidateAccessMethod({ RBACModule: 'roles' })
   async create(@Body() createDto: CreateDto, @Req() req: Request) {
-    // RBAC: Only SUPER-ADMIN can access this endpoint
-    // this.validateAccessService.check((req as any).user, [1]); // 1 = SUPER-ADMIN
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
+    }
+    if (!hasCreateAccess(req)) return denyRoleBasedAccess();
+
     const result = (await this.rolesService.create(
       RoleEntity,
       createDto,
@@ -94,7 +156,13 @@ export class RolesController {
   }
 
   @Put('updatecard')
-  async update(@Body() updateDto: UpdateDto) {
+  @ValidateAccessMethod({ RBACModule: 'roles' })
+  async update(@Body() updateDto: UpdateDto, @Req() req: Request) {
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
+    }
+    if (!hasUpdateAccess(req)) return denyRoleBasedAccess();
+
     if (!updateDto.id) {
       throw new BadRequestException('ID is required for updating');
     }
@@ -107,7 +175,13 @@ export class RolesController {
   }
 
   @Delete('deletecard')
-  async delete(@Body('id', ParseIntPipe) id: number) {
+  @ValidateAccessMethod({ RBACModule: 'roles' })
+  async delete(@Body('id', ParseIntPipe) id: number, @Req() req: Request) {
+    if (!isGranted(req)) {
+      throw new ForbiddenException(getValidationError(req) || 'Access denied');
+    }
+    if (!hasDeleteAccess(req)) return denyRoleBasedAccess();
+
     return this.rolesService.delete(RoleEntity, id);
   }
 }
